Migrate to SqlDelight 2

+++ shared/database/src/iosTest/kotlin/com.jdamcd.runlog.shared.database/ActualTestUtil.kt
@@ -1,9 +1,9 @@
 package com.jdamcd.runlog.shared.database
 
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.native.NativeSqliteDriver
+import app.cash.sqldelight.driver.native.wrapConnection
 import co.touchlab.sqliter.DatabaseConfiguration
-import com.squareup.sqldelight.db.SqlDriver
-import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
-import com.squareup.sqldelight.drivers.native.wrapConnection
 
 internal actual fun testDbDriver(): SqlDriver = NativeSqliteDriver(
     DatabaseConfiguration(
@@ -11,7 +11,7 @@ internal actual fun testDbDriver(): SqlDriver = NativeSqliteDriver(
             wrapConnection(connection) { RunLogDB.Schema.create(it) }
         },
         name = "runlog.db",
-        version = RunLogDB.Schema.version,
+        version = RunLogDB.Schema.version.toInt(),
         inMemory = true
     )
 )
